fix : 오프로드 등록 장소 리뷰 기반 수정 완료

+++ src/main/java/site/offload/offloadserver/api/place/dto/request/RegisteredPlacesRequest.java
@@ -1,10 +1,10 @@
 package site.offload.offloadserver.api.place.dto.request;
 
 public record RegisteredPlacesRequest(
-        float currentLatitude,
-        float currentLongitude
+        double currentLatitude,
+        double currentLongitude
 ) {
-    public static RegisteredPlacesRequest of(float currentLatitude, float currentLongitude) {
+    public static RegisteredPlacesRequest of(double currentLatitude, double currentLongitude) {
         return new RegisteredPlacesRequest(currentLatitude, currentLongitude);
     }
 }
