+++ src/main/java/site/offload/offloadserver/api/place/dto/response/RegisteredPlaceResponse.java
@@ -8,11 +8,11 @@ public record RegisteredPlaceResponse(
         String address,
         String shortIntroduction,
         PlaceCategory placeCategory,
-        float latitude,
-        float longitude,
-        int visitCount
+        double latitude,
+        double longitude,
+        Long visitCount
 ) {
-    public static RegisteredPlaceResponse of(Long id, String name, String address, String shortIntroduction, PlaceCategory placeCategory, float latitude, float longitude, int visitCount) {
+    public static RegisteredPlaceResponse of(Long id, String name, String address, String shortIntroduction, PlaceCategory placeCategory, double latitude, double longitude, Long visitCount) {
         return new RegisteredPlaceResponse(id, name, address, shortIntroduction, placeCategory, latitude, longitude, visitCount);
     }
 }
