+++ src/main/java/site/offload/offloadserver/db/place/repository/VisitedPlaceRepositoiry.java
@@ -8,5 +8,5 @@ import site.offload.offloadserver.db.place.entity.VisitedPlace;
 import java.util.List;
 
 public interface VisitedPlaceRepositoiry extends CrudRepository<VisitedPlace, Long> {
-    int countByMemberIdAndPlaceId(Long memberId, Long placeId);
+    Long countByMemberIdAndPlaceId(Long memberId, Long placeId);
 }
